+++ compiler/src/main/java/info/teksol/mc/mindcode/logic/mimex/ContentType.java
@@ -10,25 +10,34 @@ import java.util.stream.Stream;
 @NullMarked
 public enum ContentType {
     UNKNOWN,
-    BLOCK(true),
-    ITEM(true),
-    LIQUID(true),
-    STATUS,
-    TEAM(true),
-    UNIT(true),
-    UNIT_COMMAND,
+    ITEM(0, true),
+    BLOCK(1, true),
+    LIQUID(4, true),
+    STATUS (5),
+    UNIT(6, true),
+    WEATHER(7),
+    TEAM(15, true),
+    UNIT_COMMAND(16),
+    UNIT_STANCE(17),
     LACCESS,
     LVAR,
-    WEATHER,
     ;
 
+    public final int id;
     public final boolean hasLookup;
 
-    ContentType(boolean hasLookup) {
+    ContentType(int id, boolean hasLookup) {
+        this.id = id;
         this.hasLookup = hasLookup;
     }
 
+    ContentType(int id) {
+        this.id = id;
+        this.hasLookup = false;
+    }
+
     ContentType() {
+        this.id = -1;
         this.hasLookup = false;
     }
 
@@ -37,14 +46,25 @@ public enum ContentType {
     }
 
     private static final Map<String, ContentType> VALUE_MAP = createValueMap();
+    private static final Map<Integer, ContentType> ID_MAP = createIdMap();
 
     private static Map<String, ContentType> createValueMap() {
         return Stream.of(ContentType.values())
                 .collect(Collectors.toMap(Enum::name, e -> e));
     }
 
+    private static Map<Integer, ContentType> createIdMap() {
+        return Stream.of(ContentType.values())
+                .filter(e -> e.id >= 0)
+                .collect(Collectors.toMap(e -> e.id, e -> e));
+    }
+
     public static @Nullable ContentType byName(String contentType) {
         return VALUE_MAP.get(contentType);
+    }
+
+    public static @Nullable ContentType byId(int id) {
+        return ID_MAP.get(id);
     }
 
     public static ContentType byName(String contentType, ContentType defaultValue) {