+++ schemacode/src/main/java/info/teksol/schemacode/config/UnitOrBlockConfiguration.java
@@ -1,4 +1,0 @@
-package info.teksol.schemacode.config;
-
-public interface UnitOrBlockConfiguration extends Configuration {
-}