+++ schemacode/src/main/java/info/teksol/schemacode/mindustry/ContentConfiguration.java
@@ -1,0 +1,10 @@
+package info.teksol.schemacode.mindustry;
+
+import info.teksol.mc.mindcode.logic.mimex.ContentType;
+import info.teksol.schemacode.config.Configuration;
+
+public interface ContentConfiguration extends Configuration {
+    ContentType getContentType();
+    String getContentName();
+    int getId();
+}