+++ schemacode/src/main/java/info/teksol/schemacode/mindustry/ItemConfiguration.java
@@ -1,27 +1,38 @@
 package info.teksol.schemacode.mindustry;
 
+import info.teksol.mc.mindcode.logic.mimex.ContentType;
 import info.teksol.mc.mindcode.logic.mimex.Item;
 import info.teksol.schemacode.SchematicsMetadata;
-import info.teksol.schemacode.config.Configuration;
 
 import java.util.Objects;
 
-public final class ItemConfiguration implements Configuration {
+public final class ItemConfiguration implements ContentConfiguration {
     private final Item item;
 
     private ItemConfiguration(Item item) {
         this.item = Objects.requireNonNull(item);
     }
 
+    @Override
+    public ContentType getContentType() {
+        return ContentType.ITEM;
+    }
+
+    @Override
     public int getId() {
         return item.id();
+    }
+
+    @Override
+    public String getContentName() {
+        return item.contentName();
     }
 
     public String getName() {
         return item.name();
     }
 
-    private static ItemConfiguration forItem(Item item) {
+    public static ItemConfiguration forItem(Item item) {
         return item == null ? null : new ItemConfiguration(item);
     }