+++ schemacode/src/main/java/info/teksol/schemacode/mindustry/UnitOrBlockConfiguration.java
@@ -1,0 +1,4 @@
+package info.teksol.schemacode.mindustry;
+
+public interface UnitOrBlockConfiguration extends ContentConfiguration {
+}